// apps/api/generated/server/model/signValidatorResponse.ts
/**
 * QZD Platform API
 * The QZD Platform API exposes authentication, account, ledger, remittance, and validator administration capabilities for wallet operators and platform integrations. The contract is defined up-front to enable client/server scaffolding via OpenAPI Generator. 
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SignValidatorResponse { 
    validatorId: string;
    signature: string;
    signedAt?: string;
    /**
     * Identifier of the signing key material.
     */
    keyId?: string;
}